requirements.txt
feedparser>=6.0
requests>=2.28
trafilatura>=1.6
PyYAML>=6.0

# LLM provider packages — install whichever provider(s) you use.
# Gemini (free tier, default):
google-genai>=1.0
# Anthropic Claude:
anthropic>=0.25
# OpenAI:
openai>=1.0

# development / testing
pytest>=7.0
